make pep8-naming like _windows.py
pep8-naming only allows top-level variables to be in all caps (so as to
be constants). "INFINITE" was renamed UINT32_MAX (since that's what it
is) and the usage is clarified in source where necessary.

This is a pretty minor and debateable change just to get the style
checker to shut up.