Implement replacement of existing readers/writers
The behaviour of calling add_reader/add_writer on the same fd without
removing the reader/writer before is currently quite erratic. It is
seemingly undefined which reader/writer gets called, the old or the new,
as the previous notifier is not disabled and disconnected properly.

Note: actually I would expect that garbage collection of the notifier is
sufficient; this does not seem to be the case. It is certainly possible
though that garbage collection does not happen in time, so in any case
it is safer to disable and disconnect it beforehands.

Note: only disabling was not sufficient, for unknown reasons.

The new behaviour conforms to [PEP 3156, Section "I/O-Callbacks"][0]:

> ``add_reader(fd, callback, *args)``. Arrange for ``callback(*args)``
> to be called whenever file descriptor ``fd`` is deemed ready for
> reading. Calling ``add_reader()`` again for the same file descriptor
> implies a call to ``remove_reader()`` for the same file descriptor.

Analogous to double remove of a reader/writer, this still issues a
warning.

   [0]: https://www.python.org/dev/peps/pep-3156/#i-o-callbacks