Add strict test for scheduling of callbacks
This test shows a difference to stock asyncio event loop. In stock
asyncio, writer_cb is called exactly once, due to fut_cb being called
before writer_cb gets a second chance. writer_cb gets thus removed and
is not called again.

This may be critical. The asyncio implementation shipping with Python
3.4.2 relies on this in the implementation of sock_connect (c.f.
selector_events.py:352--389 from that version), and not adhering to this
yields to odd results up to infinite loops of callbacks.