Use type of app param to QEventLoop to determine which Qt module to use
This is a big commit, but any smaller changes would put the project into
a non-working state where everything fails and things crumble and it's
all terrible.

The big change is QtCore, and QApplication are no longer imported by
quamash at import time. Instead, the type of the app parameter to
QEventLoop determines which module is imported (PySide, PyQt4, or
PyQt5). This is fairly robust and should support PySide with Qt5 when
(and if) it is released.

QThreadExecutor now takes in a parameter that should reference
QtCore.QThread.

Changes to tests and CI configuration to use a --qtimpl parameter
instead of the QUAMASH_QTIMPL environment variable.

a new fixture called qtcore was also created so tests (mostly those that
test QThreadExecutor) could access a reference to the QtCore module.

TODO update README